Move SelectMultipleInput code to seperate file
